requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-filter==2.3.0
requests==2.24.0
djangorestframework-simplejwt==4.4.0
